# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-performance backend API for HiDocs Form \u0026 Exam Maker application using Golang and Domain-Driven Design (DDD).",
        "title": "HiDocs Backend API (Form \u0026 Exam Maker)",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "HiDocs Support",
            "email": "support@hidocs.id"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/forms": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forms"
                ],
                "summary": "List user forms ('My Forms')",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status (DRAFT, ACTIVE, CLOSED)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.FormResponseDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forms"
                ],
                "summary": "Create a new form",
                "parameters": [
                    {
                        "description": "Create Form Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateFormRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.FormResponseDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/forms/import-docx": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forms"
                ],
                "summary": "Import form from Word .docx file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Word Document (.docx)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.FormResponseDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/forms/{form_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forms"
                ],
                "summary": "Get form details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Form ID",
                        "name": "form_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.FormResponseDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forms"
                ],
                "summary": "Update form details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Form ID",
                        "name": "form_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Form Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateFormRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.FormResponseDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Forms"
                ],
                "summary": "Delete a form",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Form ID",
                        "name": "form_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/forms/{form_id}/settings": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forms"
                ],
                "summary": "Update form settings (Timer, Schedule, One-Time Submission, Randomization)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Form ID",
                        "name": "form_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Form Settings Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateFormSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.FormSettings"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/public/forms/{short_code}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Access public form by custom URL slug or short code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Custom URL Slug or Form ID",
                        "name": "short_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.PublicFormDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/public/forms/{short_code}/qr": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Generate QR Code URL for public form access",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Custom URL Slug or Form ID",
                        "name": "short_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.FormSettings": {
            "type": "object",
            "properties": {
                "auto_active_days": {
                    "type": "integer"
                },
                "duration_minutes": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "form_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active_immediately": {
                    "type": "boolean"
                },
                "is_one_time_submission": {
                    "type": "boolean"
                },
                "randomize_options": {
                    "type": "boolean"
                },
                "randomize_questions": {
                    "type": "boolean"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "domain.FormStatus": {
            "type": "string",
            "enum": [
                "DRAFT",
                "ACTIVE",
                "CLOSED"
            ],
            "x-enum-varnames": [
                "StatusDraft",
                "StatusActive",
                "StatusClosed"
            ]
        },
        "domain.FormType": {
            "type": "string",
            "enum": [
                "SURVEY",
                "EXAM"
            ],
            "x-enum-varnames": [
                "TypeSurvey",
                "TypeExam"
            ]
        },
...
        "domain.QuestionType": {
            "type": "string",
            "enum": [
                "SHORT_TEXT",
                "LONG_TEXT",
                "MULTIPLE_CHOICE",
                "CHECKBOXES",
                "DROPDOWN",
                "RATING",
                "YES_NO",
                "MATH",
                "CODE",
                "IMAGE"
            ],
            "x-enum-varnames": [
                "TypeShortText",
                "TypeLongText",
                "TypeMultipleChoice",
                "TypeCheckboxes",
                "TypeDropdown",
                "TypeRating",
                "TypeYesNo",
                "TypeMath",
                "TypeCode",
                "TypeImage"
            ]
        },
...
        "dto.CreateFormRequest": {
            "type": "object",
            "required": [
                "title",
                "type"
            ],
            "properties": {
                "custom_url": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "is_template": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                },
                "type": {
                    "enum": [
                        "SURVEY",
                        "EXAM"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.FormType"
                        }
                    ]
                }
            }
        },
...
        "dto.FormResponseDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "custom_url": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "form_settings": {
                    "$ref": "#/definitions/domain.FormSettings"
                },
                "id": {
                    "type": "string"
                },
                "is_template": {
                    "type": "boolean"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.QuestionDTO"
                    }
                },
                "response_count": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/domain.FormStatus"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/domain.FormType"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.OptionDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "is_correct": {
                    "type": "boolean"
                },
                "option_text": {
                    "type": "string"
                },
                "order_index": {
                    "type": "integer"
                },
                "question_id": {
                    "type": "string"
                }
            }
        },
        "dto.PublicFormDTO": {
            "type": "object",
            "properties": {
                "custom_url": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "form_settings": {
                    "$ref": "#/definitions/dto.PublicFormSettings"
                },
                "id": {
                    "type": "string"
                },
                "is_template": {
                    "type": "boolean"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PublicQuestionDTO"
                    }
                },
                "status": {
                    "$ref": "#/definitions/domain.FormStatus"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/domain.FormType"
                }
            }
        },
        "dto.PublicFormSettings": {
            "type": "object",
            "properties": {
                "auto_active_days": {
                    "type": "integer"
                },
                "duration_minutes": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "is_active_immediately": {
                    "type": "boolean"
                },
                "is_one_time_submission": {
                    "type": "boolean"
                },
                "randomize_options": {
                    "type": "boolean"
                },
                "randomize_questions": {
                    "type": "boolean"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "dto.PublicOptionDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "option_text": {
                    "type": "string"
                },
                "order_index": {
                    "type": "integer"
                }
            }
        },
        "dto.PublicQuestionDTO": {
            "type": "object",
            "properties": {
                "code_language": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "img_url": {
                    "type": "string"
                },
                "is_auto_scored": {
                    "type": "boolean"
                },
                "is_required": {
                    "type": "boolean"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PublicOptionDTO"
                    }
                },
                "order_index": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer"
                },
                "question_text": {
                    "type": "string"
                },
                "question_type": {
                    "$ref": "#/definitions/domain.QuestionType"
                }
            }
        },
        "dto.QuestionDTO": {
            "type": "object",
            "properties": {
                "code_language": {
                    "type": "string"
                },
                "form_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "img_url": {
                    "type": "string"
                },
                "is_auto_scored": {
                    "type": "boolean"
                },
                "is_required": {
                    "type": "boolean"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.OptionDTO"
                    }
                },
                "order_index": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer"
                },
                "question_text": {
                    "type": "string"
                },
                "question_type": {
                    "$ref": "#/definitions/domain.QuestionType"
                }
            }
        },
...
        "dto.UpdateFormRequest": {
            "type": "object",
            "required": [
                "status",
                "title",
                "type"
            ],
            "properties": {
                "custom_url": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "is_template": {
                    "type": "boolean"
                },
                "status": {
                    "enum": [
                        "DRAFT",
                        "ACTIVE",
                        "CLOSED"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.FormStatus"
                        }
                    ]
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                },
                "type": {
                    "enum": [
                        "SURVEY",
                        "EXAM"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.FormType"
                        }
                    ]
                }
            }
        },
        "dto.UpdateFormSettingsRequest": {
            "type": "object",
            "properties": {
                "auto_active_days": {
                    "type": "integer"
                },
                "duration_minutes": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "is_active_immediately": {
                    "type": "boolean"
                },
                "is_one_time_submission": {
                    "type": "boolean"
                },
                "randomize_options": {
                    "type": "boolean"
                },
                "randomize_questions": {
                    "type": "boolean"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
...
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}